refactor(common): move zstd codec out of node
